requirements.txt
colorlog==6.10.1
edge_tts==7.2.3
langchain_classic==1.0.0
langchain_community==0.4
langchain_core==1.0.1
langchain_openai==1.0.1
langgraph==1.0.1
numpy==2.3.4
opencc_python_reimplemented==0.1.7
pvporcupine==3.0.5
pyaudio==0.2.14
pydantic==2.12.3
pydub==0.25.1
pygame==2.6.1
pyside6==6.10.0
pyside6_addons==6.10.0
pyside6_essentials==6.10.0
pytest==8.4.2
python-dotenv==1.1.1
PyYAML==6.0.3
Requests==2.32.5
torch==2.9.0
transformers==4.57.1
